back/src/main/java/uno/supermercado/service/ProductoService.java
package uno.supermercado.service;

import java.security.SecureRandom;
import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import uno.supermercado.model.Producto;
import uno.supermercado.repository.ProductoRepository;

@Service
@Transactional
public class ProductoService {

    @Autowired
    private ProductoRepository productoRepository;

    @Transactional(readOnly = true)
    public List<Producto> findAll() {
        return productoRepository.findAll();
    }

    @Transactional(readOnly = true)
    public Producto findById(Long id) {
        return productoRepository.findById(id).orElse(null);
    }

    @Transactional(readOnly = true)
    public Producto save(Producto producto) {
        producto.setImagen(concatenateRandomString(producto.getImagen()));
        return productoRepository.save(producto);
    }

    public void deleteById(Long id) {
        productoRepository.deleteById(id);
    }

    private String concatenateRandomString(String imagen) {
        String randomString = generateRandomString(3);
        int dotIndex = imagen.lastIndexOf('.');
        if (dotIndex != -1) {
            return imagen.substring(0, dotIndex) + randomString + imagen.substring(dotIndex);
        } else {
            return imagen + randomString;
        }
    }

    private String generateRandomString(int length) {
        String characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
        SecureRandom random = new SecureRandom();
        StringBuilder sb = new StringBuilder(length);
        for (int i = 0; i < length; i++) {
            sb.append(characters.charAt(random.nextInt(characters.length())));
        }
        return sb.toString();
    }
}